requirements/prod.txt
Babel==1.3
Flask==0.10.1
Flask-Assets==0.8
Flask-Login==0.2.7
Flask-SQLAlchemy==1.0
Flask-Script==0.6.2
Flask-Babel==0.9
Flask-WTF==0.9.2
Jinja2==2.7
psycopg2==2.5.1
MarkupSafe==0.18
Pygments==1.6
SQLAlchemy==0.8.2
WTForms==1.0.4
Werkzeug==0.9.4
argparse==1.2.1
cssmin==0.2.0
eveapi==1.2.6
furl==0.3.5
gunicorn==17.5
hiredis==0.1.1
itsdangerous==0.23
jsmin==2.0.6
orderedmultidict==0.7.1
praw==2.1.11
python-ldap==2.4.13
pytz==2013.7
redis==2.8.0
requests==2.0.1
six==1.4.1
update-checker==0.6
webassets==0.8
wsgiref==0.1.2
wtforms-html5==0.1.3
